Refactor OpenAIIntegration object to be more generic
