+++ tests/contrib/openai/test_openai.py
@@ -108,7 +108,7 @@ class FilterOrg(TraceFilter):
 
 @pytest.fixture(scope="session")
 def mock_metrics():
-    patcher = mock.patch("ddtrace.contrib.openai._patch.get_dogstatsd_client")
+    patcher = mock.patch("ddtrace.contrib.trace_utils_llm.get_dogstatsd_client")
     DogStatsdMock = patcher.start()
     m = mock.MagicMock()
     DogStatsdMock.return_value = m
@@ -122,7 +122,7 @@ def mock_logs(scope="session"):
     Note that this fixture must be ordered BEFORE mock_tracer as it needs to patch the log writer
     before it is instantiated.
     """
-    patcher = mock.patch("ddtrace.contrib.openai._patch.V2LogWriter")
+    patcher = mock.patch("ddtrace.contrib.trace_utils_llm.V2LogWriter")
     V2LogWriterMock = patcher.start()
     m = mock.MagicMock()
     V2LogWriterMock.return_value = m
